@@ -7,33 +7,33 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^14.0.0-rc.0",
-		"@angular/common": "^14.0.0-rc.0",
-		"@angular/compiler": "^14.0.0-rc.0",
-		"@angular/core": "^14.0.0-rc.0",
-		"@angular/forms": "^14.0.0-rc.0",
-		"@angular/localize": "^14.0.0-rc.0",
-		"@angular/platform-browser": "^14.0.0-rc.0",
-		"@angular/platform-browser-dynamic": "^14.0.0-rc.0",
-		"@angular/router": "^14.0.0-rc.0",
-		"@lit/localize": "^0.11.2",
-		"lit": "^2.1.4",
-		"rxjs": "6.6.7",
-		"tslib": "^2.3.0",
-		"zone.js": "~0.11.4"
+		"@angular/animations": "15.0.4",
+		"@angular/common": "15.0.4",
+		"@angular/compiler": "15.0.4",
+		"@angular/core": "15.0.4",
+		"@angular/forms": "15.0.4",
+		"@angular/localize": "15.0.4",
+		"@angular/platform-browser": "15.0.4",
+		"@angular/platform-browser-dynamic": "15.0.4",
+		"@angular/router": "15.0.4",
+		"@lit/localize": "0.11.4",
+		"lit": "2.5.0",
+		"rxjs": "7.8.0",
+		"zone.js": "0.12.0"
 	},
 	"devDependencies": {
-		"@angular/cdk": "^14.0.0-rc.0",
-		"@angular/compiler-cli": "^14.0.0-rc.0",
-		"@ngx-playwright/test": "^0.2.3",
-		"@types/jasmine": "~3.10.0",
-		"@types/node": "^12.11.1",
-		"jasmine-core": "~4.0.0",
-		"karma": "~6.3.0",
-		"karma-chrome-launcher": "~3.1.0",
-		"karma-coverage": "~2.1.0",
-		"karma-jasmine": "~4.0.0",
-		"karma-jasmine-html-reporter": "~1.7.0",
-		"typescript": "4.6.4"
+		"@angular/cdk": "15.0.3",
+		"@angular/compiler-cli": "15.0.4",
+		"@ngx-lit/sass": "workspace:*",
+		"@ngx-playwright/test": "0.3.1",
+		"@types/jasmine": "4.3.1",
+		"@types/node": "18.11.18",
+		"jasmine-core": "4.5.0",
+		"karma": "6.4.1",
+		"karma-chrome-launcher": "3.1.1",
+		"karma-coverage": "2.2.0",
+		"karma-jasmine": "5.1.0",
+		"karma-jasmine-html-reporter": "2.0.0",
+		"typescript": "4.8.4"
 	}
 }
