@@ -13,39 +13,38 @@
 		"lint": "sn lint",
 		"test": "sn test"
 	},
-	"packageManager": "yarn@3.2.1",
+	"packageManager": "yarn@3.3.1",
 	"devDependencies": {
-		"@angular-devkit/build-angular": "^14.0.0-rc.0",
-		"@angular-eslint/builder": "^12.2.0",
-		"@angular/cdk": "^14.0.0-rc.0",
-		"@angular/common": "^14.0.0-rc.0",
-		"@angular/compiler": "^14.0.0-rc.0",
-		"@angular/compiler-cli": "^14.0.0-rc.0",
-		"@angular/core": "^14.0.0-rc.0",
-		"@angular/platform-browser": "^14.0.0-rc.0",
-		"@ngx-lit/build-angular": "workspace:*",
-		"@ngx-lit/localize": "workspace:*",
-		"@ngx-playwright/test": "^0.2.3",
-		"@snuggery/angular": "0.0.1",
-		"@snuggery/node": "^0.2.5",
-		"@snuggery/snuggery": "^0.6.5",
-		"@snuggery/yarn": "^0.2.2",
-		"@typescript-eslint/eslint-plugin": "^5.12.0",
-		"@typescript-eslint/parser": "^5.12.0",
-		"cspell-lib": "^5.18.5",
-		"eslint": "^8.9.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-import-resolver-node": "^0.3.6",
-		"eslint-plugin-import": "^2.25.4",
-		"express-check-in": "^0.1.0",
-		"husky": "^7.0.4",
-		"is-ci": "^3.0.1",
-		"prettier": "^2.5.1",
-		"rxjs": "6.6.7",
-		"typescript": "4.6.4",
-		"zone.js": "0.11.4"
+		"@angular-devkit/build-angular": "15.0.4",
+		"@angular-eslint/builder": "15.1.0",
+		"@angular/cdk": "15.0.3",
+		"@angular/common": "15.0.4",
+		"@angular/compiler": "15.0.4",
+		"@angular/compiler-cli": "15.0.4",
+		"@angular/core": "15.0.4",
+		"@angular/platform-browser": "15.0.4",
+		"@ngx-playwright/test": "0.3.1",
+		"@snuggery/angular": "0.3.2",
+		"@snuggery/node": "0.5.2",
+		"@snuggery/snuggery": "0.9.2",
+		"@snuggery/yarn": "0.4.2",
+		"@typescript-eslint/eslint-plugin": "5.47.1",
+		"@typescript-eslint/parser": "5.47.1",
+		"cspell-lib": "6.18.0",
+		"eslint": "8.30.0",
+		"eslint-config-prettier": "8.5.0",
+		"eslint-import-resolver-node": "0.3.6",
+		"eslint-plugin-import": "2.26.0",
+		"express-check-in": "0.1.1",
+		"husky": "8.0.2",
+		"is-ci": "3.0.1",
+		"prettier": "2.8.1",
+		"rxjs": "7.8.0",
+		"typescript": "4.8.4",
+		"zone.js": "0.12.0"
 	},
 	"resolutions": {
-		"esbuild@0.13.8": "npm:esbuild-wasm@0.13.8"
+		"resolve.exports@^1.1.0": "patch:resolve.exports@npm%3A1.1.0#./.yarn/patches/resolve.exports-npm-1.1.0-81756e03ba.patch",
+		"resolve.exports@1.1.0": "patch:resolve.exports@npm%3A1.1.0#./.yarn/patches/resolve.exports-npm-1.1.0-81756e03ba.patch"
 	}
 }
