@@ -1,46 +1,58 @@
-import {createBuilder} from '@angular-devkit/architect';
-import type {
-	BrowserBuilderOptions,
-	DevServerBuilderOptions,
-	ExtractI18nBuilderOptions,
-} from '@angular-devkit/build-angular';
+import {
+	type BuilderContext,
+	type BuilderOutput,
+	createBuilder,
+} from '@angular-devkit/architect';
+import type {BrowserBuilderOptions} from '@angular-devkit/build-angular';
 import type {JsonObject} from '@angular-devkit/core';
-import {from} from 'rxjs';
-import {switchMap} from 'rxjs/operators';
+import {from, Observable, type ObservableInput} from 'rxjs';
 
 import {
 	executeBrowserBuilder,
 	executeDevServerBuilder,
 	executeExtractI18nBuilder,
+	executeKarmaBuilder,
 } from './builders.js';
 
-export const browser = createBuilder<BrowserBuilderOptions & JsonObject>(
-	(options: BrowserBuilderOptions & JsonObject, context) =>
-		from(
-			context.validateOptions<BrowserBuilderOptions & JsonObject>(
-				options,
-				'@angular-devkit/build-angular:browser',
-			),
-		).pipe(switchMap(options => executeBrowserBuilder(options, context))),
-);
+const wrapped = new Map<
+	string,
+	(
+		// eslint-disable-next-line @typescript-eslint/no-explicit-any
+		options: any,
+		context: BuilderContext,
+	) => ObservableInput<BuilderOutput>
+>([
+	['browser', executeBrowserBuilder],
+	['dev-server', executeDevServerBuilder],
+	['extract-i18n', executeExtractI18nBuilder],
+	['karma', executeKarmaBuilder],
+]);
 
-export const devServer = createBuilder<DevServerBuilderOptions & JsonObject>(
-	(options: DevServerBuilderOptions & JsonObject, context) =>
-		from(
-			context.validateOptions<DevServerBuilderOptions & JsonObject>(
-				options,
-				'@angular-devkit/build-angular:dev-server',
-			),
-		).pipe(switchMap(options => executeDevServerBuilder(options, context))),
-);
+export default createBuilder((options, context) => {
+	const name = context.builder.builderName;
+	const builder = wrapped.get(name.includes(':') ? name.split(':')[1]! : name);
 
-export const extractI18n = createBuilder<
-	ExtractI18nBuilderOptions & JsonObject
->((options: ExtractI18nBuilderOptions & JsonObject, context) =>
-	from(
-		context.validateOptions<ExtractI18nBuilderOptions & JsonObject>(
-			options,
-			'@angular-devkit/build-angular:extract-i18n',
-		),
-	).pipe(switchMap(options => executeExtractI18nBuilder(options, context))),
-);
+	if (builder == null) {
+		throw new Error(`Unexpected builder ${name}`);
+	}
+
+	return new Observable(subscriber => {
+		context
+			.validateOptions<BrowserBuilderOptions & JsonObject>(
+				options,
+				`@angular-devkit/build-angular:${name}`,
+			)
+			.then(
+				validatedOptions => {
+					if (subscriber.closed) {
+						return;
+					}
+
+					subscriber.add(
+						from(builder(validatedOptions, context)).subscribe(subscriber),
+					);
+				},
+				error => subscriber.error(error),
+			);
+	});
+});
